requirements.txt
Flask==3.0.3
joblib==1.4.2
numpy==2.1.2
requests==2.32.3
scikit-learn==1.4.2
gunicorn==23.0.0
python-dotenv==1.0.1